fixed stuff?

+++ philo/time.c
@@ -14,17 +14,17 @@
 
 bool	time_add(struct timeval *t1, struct timeval *t2)
 {
-	if (UINT_MAX - (unsigned)t1->tv_sec < (unsigned)t2->tv_sec || (UINT_MAX
-			- (unsigned)t1->tv_sec == (unsigned)t2->tv_sec
-			&& ((unsigned)t1->tv_usec + (unsigned)t2->tv_usec) / 1000000 == 1))
+	if (UINT_MAX - t1->tv_sec < t2->tv_sec || (UINT_MAX
+			- t1->tv_sec == t2->tv_sec
+			&& (t1->tv_usec + t2->tv_usec) / 1000000 == 1))
 	{
 		t1->tv_sec = 0;
 		t1->tv_usec = 0;
 		return (true);
 	}
-	t1->tv_sec = (unsigned)t1->tv_sec + (unsigned)t2->tv_sec
-		+ ((unsigned)t1->tv_usec + (unsigned)t2->tv_usec) / 1000000;
-	t1->tv_usec = ((unsigned)t1->tv_usec + (unsigned)t2->tv_usec) % 1000000;
+	t1->tv_sec = t1->tv_sec + t2->tv_sec
+		+ (t1->tv_usec + t2->tv_usec) / 1000000;
+	t1->tv_usec = (t1->tv_usec + t2->tv_usec) % 1000000;
 	return (false);
 }
 
@@ -33,12 +33,12 @@ struct timeval	*time_sub(struct timeval *t1, struct timeval *t2)
 	if ((t1->tv_sec < t2->tv_sec) || (t1->tv_sec == t2->tv_sec
 			&& t1->tv_usec < t2->tv_usec))
 		return (NULL);
-	t1->tv_sec = (unsigned)t1->tv_sec - (unsigned)t2->tv_sec;
+	t1->tv_sec = t1->tv_sec - t2->tv_sec;
 	if (t1->tv_usec >= t2->tv_usec)
-		t1->tv_usec = (unsigned)t1->tv_usec - (unsigned)t2->tv_usec;
+		t1->tv_usec = t1->tv_usec - t2->tv_usec;
 	else
 	{
-		t1->tv_usec = 1000000 + (unsigned)t1->tv_usec - (unsigned)t2->tv_usec;
+		t1->tv_usec = 1000000 + t1->tv_usec - t2->tv_usec;
 		t1->tv_sec--;
 	}
 	return (t1);
